+++ src/components/Header.js
@@ -11,11 +11,11 @@ import logo from "../assets/logo.png";
 const Header = () => {
 
     return (
-        <nav class="bg-[#1A1818]">
-            <div className='fixed top-0 right-0 left-0 bg-gradient z-50 bg-[#1A1818]'>
+        <nav class="bg-[#008080]">
+            <div className='fixed top-0 right-0 left-0 bg-gradient z-50 bg-[#008080]'>
                 <div className='flex gap-4 justify-center items-center mx-4 pt-2'>
                    
-                <Link to="/"><img src={logo} alt="logo" class='w-40' /></Link>
+                <Link to="/"><img src={logo} alt="logo" class='w-20' /></Link>
                     <ul className='flex m-auto gap-20'>
                         <Link to='/explore' style={{ textDecoration: 'none', color: 'white' }}>Explore</Link>
                         <Link to='/dashboard' style={{ textDecoration: 'none', color: 'white' }}>Dashboard</Link>
